pyproject.toml
[project]
name = "robotframework-slacknotification"
version = "0.1.2"
description = "Send real-time notifications to Slack with the results of your tests"
authors = [
    {name = "Robot Courses",email = "git@robotcourses.com.br"}
]
license = {text = "Apache-2.0 license"}
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "slack-sdk (>=3.35.0,<4.0.0)",
    "robotframework (>=7.0,<8.0)"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.8",
    "Framework :: Robot Framework",
    "Framework :: Robot Framework :: Library",
]

[tool.poetry]
packages = [
    {include = "RobotSlackNotification"}
]

[tool.poetry.urls]
"Github" = "https://github.com/robotcourses/robotframework-slacknotification"
"Bugs Tracker" = "https://github.com/robotcourses/robotframework-slacknotification/issues"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
